requirement.txt
setuptools
keyboard
websockets
loguru
PySide6
psutil
opencv-python
plyvel-win32